# docs/swagger.yaml
definitions:
  wallet.Err:
    properties:
      message:
        type: string
    type: object
  wallet.Wallet:
    properties:
      balance:
        example: 100
        type: number
      created_at:
        example: "2024-03-25T14:19:00.729237Z"
        type: string
      id:
        example: 1
        type: integer
      user_id:
        example: 1
        type: integer
      user_name:
        example: John Doe
        type: string
      wallet_name:
        example: John's Wallet
        type: string
      wallet_type:
        example: Create Card
        type: string
    type: object
host: localhost:1323
info:
  contact: {}
  description: Sophisticated Wallet API
  title: Wallet API
  version: "1.0"
paths:
  /api/v1/user/{id}/wallets:
    delete:
      consumes:
      - application/json
      description: Delete  wallets  by user id
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wallet.Wallet'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/wallet.Err'
      summary: Delete  wallets by user id
      tags:
      - wallet
    get:
      consumes:
      - application/json
      description: Get  wallets  by user id
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wallet.Wallet'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/wallet.Err'
      summary: Get  wallets by user id
      tags:
      - wallet
  /api/v1/wallets:
    get:
      consumes:
      - application/json
      description: Get all wallets
      parameters:
      - description: Wallet type  fillter
        in: query
        name: wallet_type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wallet.Wallet'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/wallet.Err'
      summary: Get all wallets
      tags:
      - wallet
    post:
      consumes:
      - application/json
      description: Create  wallets
      parameters:
      - description: Wallet object
        in: body
        name: wallet
        required: true
        schema:
          $ref: '#/definitions/wallet.Wallet'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wallet.Wallet'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/wallet.Err'
      summary: Create  wallets
      tags:
      - wallet
    put:
      consumes:
      - application/json
      description: UPDATE Wallets by id
      parameters:
      - description: Wallet object
        in: body
        name: wallet
        required: true
        schema:
          $ref: '#/definitions/wallet.Wallet'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wallet.Wallet'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/wallet.Err'
      summary: UPDATE Wallets by id
      tags:
      - wallet
swagger: "2.0"